Move file to correct folder
