+++ lib/music_player/music_player_screen.dart
@@ -8,7 +8,7 @@ import 'package:musbx/music_player/pitch_speed_card/pitch_speed_card.dart';
 import 'package:musbx/music_player/position_card/button_panel.dart';
 import 'package:musbx/music_player/current_song_card/current_song_panel.dart';
 import 'package:musbx/music_player/position_card/position_slider.dart';
-import 'package:musbx/music_player/position_card/song_history_list.dart';
+import 'package:musbx/music_player/current_song_card/song_history_list.dart';
 import 'package:musbx/screen/card_list.dart';
 import 'package:musbx/screen/default_app_bar.dart';
 import 'package:musbx/screen/empty_tab_bar.dart';
